Conflicts:
	app/Vendor/dompdf/lib/fonts/log.htm

+++ app/Plugin/Sales/Model/ProductSpecificationProcess.php
@@ -53,7 +53,8 @@ class ProductSpecificationProcess extends AppModel {
 
 		$this->bind(array('Sales.ProductSpecificationProcessHolder'));
 		$Ids = array();
-		
+		if (!empty($processdata[$this->name])) {
+
 		foreach ($processdata[$this->name] as $key1 => $processList) {
 			
 			$this->create();
@@ -63,9 +64,9 @@ class ProductSpecificationProcess extends AppModel {
 			$processList['product_specification_id'] = $specId;
 			$processList['product_id'] = $processdata['Product']['id'];
 			$this->save($processList);
-
+			
 			array_push($Ids, $this->id.'-'.$processList['order'].'-'.'Process');
-
+			
 			unset($processdata[$this->name][$key1]['order']);
 			
 			$holder = 'ProductSpecificationProcessHolder';
@@ -90,16 +91,23 @@ class ProductSpecificationProcess extends AppModel {
 			//$this->ProductSpecificationProcessHolder->saveProcessHolder($processdata,$this->id,$auth);
 			
 		}
-		
-		return $Ids;
+	}
+	
+	return $Ids;
+
 
 	}
 
 	public function deleteData($processData = null){
 
-		foreach ($processData['ProductSpecificationProcess'] as $key => $deleteMe) {
-			$this->delete($deleteMe);
+		if (!empty($processData['ProductSpecificationProcess'] )) {
+
+			foreach ($processData['ProductSpecificationProcess'] as $key => $deleteMe) {
+				$this->delete($deleteMe);
+			}
+
 		}
+		
 
 	}
 }
